pinaxsite_project/requirements/project.txt
# project.txt is a pip requirements file which describes the distributions
# required by your project to run.

--requirement=base.txt

creole==1.1
PIL==1.1.7
psycopg2==2.3.2
Pygments==1.3.1
biblion==0.1.dev10
django-atomformat==0.1.1.dev1
django-tagging==0.3.1
django-uni-form==0.7.0
dregni==0.1.dev1
template_utils==0.4p2
nashvegas==0.2.1
httplib2==0.6.0